Split factories file into module with directory
